KedaTestScaler/requirements.txt
requests==2.32.3
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
python-json-logger==2.0.7 